+++ fabric/changelog.md
@@ -22,3 +22,4 @@
 - Added Sound Physics compatibility
 - Added optional port for voice_host setting
 - Fixed audio randomly popping
+- Added Cloth Config GUI support